Restructuring Auth endpoints
Changed the following:
 - Moved Logout to Auth Delete
 - Renamed Auth.GoogleToken to Auth.Create with more specific parameter namtes
 - Moved Oauth2 struct to it's own model (skinny controller)
 - Pulled Auth.Index into Home.Index (separation of concerns)
 - Removed Index.html in controllers folder
 - Used $.ajax for signIn and signOut methods and filter to append csrf token to specific ajax calls (dry)
 -
