+++ app/controllers/auth.go
@@ -2,61 +2,17 @@ package controllers
 
 import (
 	"github.com/revel/revel"
-	"google.golang.org/api/oauth2/v2"
-	"net/http"
-	"io/ioutil"
+	"www-forecast/app/models"
 	"fmt"
-	"encoding/json"
 )
-
-type Oauth2 struct {
-	Azp           string `json:"azp"`
-	Aud           string `json:"aud"`
-	Sub           string `json:"sub"`
-	Hd            string `json:"hd"`
-	Email         string `json:"email"`
-	EmailVerified string `json:"email_verified"`
-	AtHash        string `json:"at_hash"`
-	Exp           string `json:"exp"`
-	Iss           string `json:"iss"`
-	Jti           string `json:"jti"`
-	Iat           string `json:"iat"`
-	Name          string `json:"name"`
-	Picture       string `json:"picture"`
-	GivenName     string `json:"given_name"`
-	FamilyName    string `json:"family_name"`
-	Locale        string `json:"locale"`
-	Alg           string `json:"alg"`
-	Kid           string `json:"kid"`
-}
 
 type Auth struct {
 	*revel.Controller
 }
 
-var httpClient = &http.Client{}
+func (c Auth) Create(googleIdToken string) revel.Result {
+	ti, hd, err := models.VerifyIdToken(googleIdToken)
 
-func (c Auth) Index() revel.Result {
-
-	//c.Validation.Required(c.Session["user"])
-	//c.Validation.Required(c.Session["hd"])
-
-	//if c.Validation.HasErrors() {
-	//	fmt.Println("We had a validation error")
-		//c.Flash.Success("Please log in.")
-	//}
-
-	return c.Render()
-}
-
-func (c Auth) GoogleToken(idtoken string) revel.Result {
-	//fmt.Println("Received: ", idtoken)
-
-	ti, hd, err := verifyIdToken(idtoken)
-
-	//fmt.Println("verified email: ", ti.Email)
-	//fmt.Println("verified domain: ", hd)
-	//fmt.Printf("%+v\n", ti)
 	if (err != nil){
 		c.Flash.Error("Could not verify token from Google.")
 		fmt.Println("Can't verify hd claim from Google: " + err.Error())
@@ -78,57 +34,14 @@ func (c Auth) GoogleToken(idtoken string) revel.Result {
 	}
 
 	res := JSONResponse{Code: "ok"}
-
 	return c.RenderJSON(res)
 }
 
-func verifyIdToken(idToken string) (*oauth2.Tokeninfo, string, error) {
-    oauth2Service, err := oauth2.New(httpClient)
+func (c Auth) Delete() revel.Result {
+	c.Session["user"] = ""
+	c.Session["hd"]		= ""
+	c.Flash.Success("Logged Out.")
 
-		if err != nil {
-			fmt.Println("Cannot create HTTP client: " + err.Error())
-		}
-
-    tokenInfoCall := oauth2Service.Tokeninfo()
-    tokenInfoCall.IdToken(idToken)
-    tokenInfo, err := tokenInfoCall.Do()
-
-		if err != nil {
-			fmt.Println("Cannot get OAuth data: " + err.Error())
-		}
-
-		hd, err := getHd(idToken)
-
-    return tokenInfo, hd, err
-}
-
-//This function exists because of a shortcoming in the Google oauth2 api. `hd` is not returned in verifyIdToken(). Causes a second request to google on login.
-func getHd(idToken string) (hd string, err error) {
-	client := &http.Client{}
-	transport := &http.Transport{Proxy: http.ProxyFromEnvironment}
-	transport.DisableCompression = true
-	client.Transport = transport
-	req, err := http.NewRequest("GET", "https://www.googleapis.com/oauth2/v3/tokeninfo?id_token=" + idToken, nil)
-  if err != nil {
-			 fmt.Println("HTTP error crafting Google API Token Request" + err.Error())
-  }
-	res, err:= client.Do(req)
-	defer res.Body.Close()
-
-	if err != nil {
-			 fmt.Println("HTTP error getting Google API ID Token" + err.Error())
-	}
-
-	//fmt.Println(res)
-
-	body, err := ioutil.ReadAll(res.Body)
-
-	o := Oauth2{}
-
-	err = json.Unmarshal(body, &o)
-	if err != nil {
-			 fmt.Println("Error with JSON unmarshal" + err.Error())
-  }
-	return o.Hd, err
-
+	res := JSONResponse{Code: "ok"}
+	return c.RenderJSON(res)
 }
