+++ app/controllers/home.go
@@ -1,0 +1,13 @@
+package controllers
+
+import (
+	"github.com/revel/revel"
+)
+
+type Home struct {
+	*revel.Controller
+}
+
+func (c Home) Index() revel.Result {
+	return c.Render()
+}
